myapp/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
mysql-connector-python
passlib
python-jose